Merge of the methods onHandleIntent and the second startService
There was no valuable argument for the separation of thoses two methods
wich were both supposed to be the entry point of the service.
On top of that the name of the second method was identical to a public static method
which made it even more confusing.
non static startService code have been therefore moved into onHandleIntent.