Adds fields in TestInfo, change encapsulation and usage
Adds 4 fields (frameDropped, percentOfBadScreenshots, percentOfBadSeek, numberOfWarnings);
Adds associated getters for said fields.
Change the accessibility of its fields to protected so that in-package class could change
its fields directly.
Change its constructor to proctected for the same reason as above and remove Score argument
as it's now created by the constructor.
Update usage of the class in BenchService's function testFile.