Improve cleanup process of BenchService & BenchServiceDispatcher
Make sure that once a DONE_STATUS has been received we stop/unbind the
service and clean the private fields of the dispatcher.
As a consequence we prevent call of statService is an other one has been
launched by the same instance of BenchServiceDispatcher.
Also in the BenchSerive's sendMessage method after sending the done event
set the dispatcher field to null.