Improve encapsulation and correct misunderstanding
In the commit 'Adds fields in TestInfo, change encapsulation and usage'
we swiched TestInfo specifier to final and its fields to proctected in the
hope that it will allow only this class and the other one in the same package
to use its fields and constructor.
The way we used was an overkill and is not very clear.
We now instead remove the final specifier and all the protected ones also as
something with no specifier is only visible by the class and its package (as
 we whised).