New tool class ScreenshotValidator to check the validity of the screenshots
This class contains only public static methods to use to get the color difference
between the average of an image saved inside a int in RGB and a screenshot file.
To do so call getValidityPercent.
It find the average of the given image file, translate the rgb colors into
lab color space and use the CIE94 algorithm to find the color difference.