Update MediaInfo internals, snapshots split into two lists
The snapshots member in MediaInfo was, until then, an Array of two
cases holding two ArrayList<Long>.
The second case shouldn't be a ArrayList<Long> but an ArrayList<Integer> so we
update the code accordingly.

We then update the JSonParser class as its the class responsible for creating
the instances of MediaInfo.