Correct issue where if an error occured in Service it would never open again
If an error occurs in BenchService we do not set to null the dispatcher field and
in BenchServiceDispatcher we set the initContext to null.
When calling startService we do not explicitly bind with the service if the service
connection is still present (occurs only if an error occured in service).