Now handles when VLC is killed
If VLC's BenchActivity is killed after we launched it
we will receive a null Intent along with a 0 result code.
Using this we made so that if thoses conditions are met we
read a specific public shared preference of VLC to get the
error message on which VLC crashed and display it inside a popup (temporarily).
We then continue to the next test.