Remove static resumeMedia for error during download
The resumeMedia field was set if an error occured and
during download/sha512 process jumped all the media already
checked and tried directly from the one held in resumtMedia.
This is has been found to be a source of bugs and is not
that much of a gain of time (as the sha512 is really fast).