Stop using vlc's error codes as a fatal errors
Until then if vlc's BenchActivity finished itself and set its result
with a code different from -1 (RESULT_OK) we stopped the whole testing
process and displayed a dialog.

We now handle it as we do for VLC crashes, that is to say:
we display the error message in a dialog, set the test as failed in
TestPage's TextView for logging and move on to the next test once
the user clicked the dialog.