Now displays the download bitrate while donwloading videos
  Adds a second argument of type long in BenchServiceListener's updatePercent
to receive the current download bitrate if need be.

  When downloading BenchService computes its download bit rate and send it
along with the new progress inside a Pair<Double, Long>
to BenchServiceDispatcher with the new status PERCENT_STATUS_BITRATE.

  When BenchServiceDispatcher receives said status it send the content of
the received Pair<Double, Long> to its listeners through the method updatePercent.
If it, instead, receives a PERCENT_STATUS status it will call the same method
on its listeners but with a the bitrate value of -1 (BenchServiceDispatcher.NO_BITRATE).

  TestPage's updatePercent adds this new argument to its signature. It checks for
BencserviceDispatcher.NO_BITRATE on said argument, if it's the case update the view
as before this commit.
  Otherwise it uses a new string format to display update the view with the progress and
the  bitrate.
  To obtain the string representation of the bitrate it calls the new method bitRateToString.