Split MainPage into a child and father to separate logic and UI.
This moves everything not related to UI from MainPage to VLCWorkerModel
and make MainPage extends VLCWorkerModel.

It allows to change the View/UI independently from the core of the
application.

To do so the new VLCWorkerModel class contains a number of abstract method that it
calls when UI should/could be updated.

On MainPage's side we override all thoses methods and tranfer all the UI code
into it.