For ResultPage one listener for all items instead of one for each.
Before this every row of every tab of ResultPage had its own
click listener.
Now each tab has one listener only and it receives an index to
know for what row should it should create a new Dialog.

The result is that the content of the listener has been moved from ResultAdapter
to GridFragment.