Adds way for BenchService to send String once a step is finished to listeners.
Adds method stepFinished(String) in BenchServiceListener.

Make MainPage implement it an append the string to its TextView used for logging.

Adds a constant in BenchService so that it can be called by BenchService to send it
to BenchServiceDispatcher along with an object.

Make BenchServiceDispatcher use that constant to call stepFinished on all its listeners
with the string received from BenchService.