Create a tool activity to automatically receive the sha512 and colors of video
this activity is only meant for dev/debug usage,
It allows us to give the name of a new video added on our media server
and to give it some timestamps.
Once finished the activity will have calculated the sha512 of the video and
the average color at the given moments in a Dialog and in the logs.