fix restarting vlc too fast
This is a temporary workaround a deeper problem.
When a quality test fails, there is no screenshot analysis,
hence no delay before re-starting vlc-android.
That sometimes doesn't let vlc-android exit properly before
being recreated. For that specific case a delay was introduced.
But the class should be refactored as to not start anything
in onActivityResult() but rather later, in onResume() for example.