Refactor: remove MainPageDownloadFragment
Download and checks are now performed at
the start of a benchmark. This simplifies
handling the state of the app (always
checking if the samples were checked
or downloaded). Furthermore, this prepares
the next refactors, using the new progress
dialog, and creating a specific class to
handle the benchmark instead of
VLCWorkerModel